[ig-colors] Update regression stories
